Refactored as genearl purpose DebugMode
